chore: migrate to `tsdown` for bundling
