+++ .changeset/red-mails-begin.md
@@ -1,0 +1,13 @@
+---
+"@cf-wasm/minify-html": patch
+"@cf-wasm/internals": patch
+"@cf-wasm/plugins": patch
+"@cf-wasm/quickjs": patch
+"@cf-wasm/photon": patch
+"@cf-wasm/satori": patch
+"@cf-wasm/resvg": patch
+"@cf-wasm/png": patch
+"@cf-wasm/og": patch
+---
+
+chore: migrate to `tsdown` for bundling
