+++ packages/internals/package.json
@@ -22,8 +22,8 @@
     "lint": "biome lint ./src",
     "check": "biome check ./src",
     "check:types": "tsc --noEmit",
-    "build": "NODE_ENV=production tsup",
-    "dev": "NODE_ENV=development tsup --watch"
+    "build": "NODE_ENV=production tsdown",
+    "dev": "NODE_ENV=development tsdown --watch"
   },
   "files": [
     "dist"
@@ -67,6 +67,6 @@
   },
   "devDependencies": {
     "@types/node": "catalog:default",
-    "tsup": "catalog:default"
+    "tsdown": "catalog:default"
   }
 }
