+++ packages/minify-html/src/edge-light.ts
@@ -3,5 +3,5 @@ import { initMinifyHTML } from './minify-html';
 
 initMinifyHTML.sync(minifyHTMLWasmModule);
 
-export { minifyHTMLWasmModule };
 export * from './minify-html';
+export { minifyHTMLWasmModule };
