+++ packages/og/tsdown.config.ts
@@ -1,15 +1,9 @@
 import fs from 'node:fs';
 import path from 'node:path';
-import { fileURLToPath } from 'node:url';
 import * as glob from 'glob';
-import { defineConfig, type Options } from 'tsup';
-
-const YOGA_WASM_LOCATION = path.join(fileURLToPath(import.meta.resolve('satori')), '../../yoga.wasm');
-const YOGA_WASM_DESTINATION = 'src/lib/yoga.wasm';
+import { defineConfig, type UserConfig } from 'tsdown';
 
 export default defineConfig(() => {
-  fs.copyFileSync(YOGA_WASM_LOCATION, YOGA_WASM_DESTINATION);
-
   // generate inline modules
   for (const file of glob.sync('src/**/*.{wasm,bin,txt}')) {
     const content = fs.readFileSync(file);
@@ -29,20 +23,34 @@ export default defineConfig(() => {
   const commonOptions = {
     outDir: 'dist',
     platform: 'neutral',
+    target: 'es2018',
     sourcemap: true,
-    splitting: true,
-    bundle: true,
-    skipNodeModulesBundle: true,
+    unbundle: true,
+    deps: {
+      skipNodeModulesBundle: true,
+    },
     shims: true,
     dts: true,
-  } satisfies Options;
+    ignoreWatch: ['.turbo'],
+  } satisfies UserConfig;
 
   return [
     {
       ...commonOptions,
-      entry: ['src/edge-light.ts', 'src/node.ts', 'src/others.ts', 'src/workerd.ts'],
+      entry: [
+        'src/edge-light.ts',
+        'src/node.ts',
+        'src/others.ts',
+        'src/workerd.ts',
+        'src/figma.ts',
+        'src/html-to-react.ts',
+        'src/lib/**/*.{js,d.ts}',
+      ],
       format: ['esm'],
-      external: [/\.wasm$/, /\.wasm\?module$/, /\.bin$/, /\.txt$/],
+      deps: {
+        ...commonOptions.deps,
+        neverBundle: [/\.wasm$/, /\.wasm\?module$/, /\.bin$/, /\.txt$/],
+      },
       clean: true,
       async onSuccess() {
         // Copy assets
@@ -62,8 +70,8 @@ export default defineConfig(() => {
     },
     {
       ...commonOptions,
-      entry: ['src/node.ts', 'src/others.ts'],
+      entry: ['src/node.ts', 'src/others.ts', 'src/figma.ts', 'src/html-to-react.ts'],
       format: ['cjs'],
     },
-  ];
+  ] satisfies UserConfig[];
 });
