+++ packages/plugins/package.json
@@ -22,8 +22,8 @@
     "lint": "biome lint ./src",
     "check": "biome check ./src",
     "check:types": "tsc --noEmit",
-    "build": "NODE_ENV=production tsup",
-    "dev": "NODE_ENV=development tsup --watch"
+    "build": "NODE_ENV=production tsdown",
+    "dev": "NODE_ENV=development tsdown --watch"
   },
   "files": [
     "dist"
@@ -75,16 +75,6 @@
       }
     }
   },
-  "dependencies": {
-    "magic-string": "^0.30.21"
-  },
-  "devDependencies": {
-    "@types/node": "catalog:default",
-    "nitro": "3.0.1-alpha.1",
-    "rollup": "^4.53.3",
-    "tsup": "catalog:default",
-    "vite": "catalog:default"
-  },
   "peerDependencies": {
     "rollup": "^4.0.0",
     "vite": "^6.1.0 || ^7.0.0"
@@ -96,5 +86,15 @@
     "vite": {
       "optional": true
     }
+  },
+  "dependencies": {
+    "magic-string": "^0.30.21"
+  },
+  "devDependencies": {
+    "@types/node": "catalog:default",
+    "nitro": "3.0.260415-beta",
+    "rollup": "^4.60.2",
+    "tsdown": "catalog:default",
+    "vite": "catalog:default"
   }
 }
