+++ packages/png/src/workerd.ts
@@ -3,5 +3,5 @@ import { initPng } from './png';
 
 initPng.sync(pngWasmModule);
 
-export { pngWasmModule };
 export * from './png';
+export { pngWasmModule };
