+++ packages/quickjs/package.json
@@ -18,8 +18,8 @@
     "lint": "biome lint ./src",
     "check": "biome check ./src",
     "check:types": "tsc --noEmit",
-    "build": "NODE_ENV=production tsup",
-    "dev": "NODE_ENV=development tsup --watch"
+    "build": "NODE_ENV=production tsdown",
+    "dev": "NODE_ENV=development tsdown --watch"
   },
   "files": [
     "dist"
@@ -125,6 +125,6 @@
   "devDependencies": {
     "@types/node": "catalog:default",
     "glob": "catalog:default",
-    "tsup": "catalog:default"
+    "tsdown": "catalog:default"
   }
 }
