+++ packages/quickjs/src/node.ts
@@ -14,5 +14,5 @@ export async function getQuickJSWASMModule() {
   return singletonPromise;
 }
 
-export { releaseSyncWasmModule };
 export * from './lib/release';
+export { releaseSyncWasmModule };
