+++ packages/quickjs/src/workerd-debug.ts
@@ -14,5 +14,5 @@ export async function getQuickJSWASMModule() {
   return singletonPromise;
 }
 
-export { debugSyncWasmModule, debugSyncWasmModuleSourceMap };
 export * from './lib/debug';
+export { debugSyncWasmModule, debugSyncWasmModuleSourceMap };
