+++ packages/satori/src/edge-light.ts
@@ -4,5 +4,5 @@ import { initSatori, satori } from './satori';
 initSatori(yogaWasmModule);
 
 export default satori;
-export { yogaWasmModule };
 export * from './satori';
+export { yogaWasmModule };
