+++ packages/satori/src/node.ts
@@ -6,5 +6,5 @@ const yogaWasmModule = new WebAssembly.Module(yogaWasmBuffer);
 initSatori(yogaWasmModule);
 
 export default satori;
-export { yogaWasmModule };
 export * from './satori';
+export { yogaWasmModule };
