+++ pnpm-workspace.yaml
@@ -1,17 +1,23 @@
-{
-  "packages": ["packages/*", "examples/*", "__tests__", "__tests__/*"],
-  "onlyBuiltDependencies":
-    ["@biomejs/biome", "esbuild", "sharp", "wasm-pack", "workerd"],
-  "catalog":
-    {
-      "@types/node": "^24.10.1",
-      "@types/react": "^19.2.3",
-      "glob": "^11.0.3",
-      "react": "^19.2.0",
-      "tsup": "^8.5.0",
-      "vite": "^7.2.2",
-      "vitest": "^4.0.8",
-      "wasm-pack": "^0.13.1",
-      "wrangler": "^4.47.0",
-    },
-}
+packages:
+  - packages/*
+  - examples/*
+  - __tests__
+  - __tests__/*
+
+catalog:
+  "@types/node": ^25.6.0
+  "@types/react": ^19.2.14
+  glob: ^11.0.3
+  react: ^19.2.5
+  tsdown: ^0.21.9
+  vite: ^7.2.2
+  vitest: ^4.1.4
+  wasm-pack: ^0.13.1
+  wrangler: ^4.83.0
+
+allowBuilds:
+  "@biomejs/biome": true
+  esbuild: true
+  sharp: true
+  wasm-pack: true
+  workerd: true
